Move checkconfig package to internal/ and update check.go
