+++ internal/checkconfig/checkconfig.go
